requirements.txt
PySide6
pytesseract
argostranslate
mss
keyboard
pyttsx3
pyperclip
python-docx

pyautogui
screeninfo
pillow
